requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
aiofiles==23.2.1
    # via aiogram
aiogram==3.8.0
    # via -r requirements.in
aiohttp==3.9.5
    # via aiogram
aiosignal==1.3.1
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
attrs==23.2.0
    # via aiohttp
beautifulsoup4==4.12.3
    # via -r requirements.in
cachetools==5.3.3
    # via google-auth
certifi==2024.7.4
    # via
    #   aiogram
    #   requests
charset-normalizer==3.3.2
    # via requests
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
google-auth==2.31.0
    # via
    #   google-auth-oauthlib
    #   gspread
google-auth-oauthlib==1.2.0
    # via gspread
gspread==6.1.2
    # via -r requirements.in
idna==3.7
    # via
    #   requests
    #   yarl
lxml==5.2.2
    # via -r requirements.in
magic-filter==1.0.12
    # via aiogram
multidict==6.0.5
    # via
    #   aiohttp
    #   yarl
oauthlib==3.2.2
    # via requests-oauthlib
pyasn1==0.6.0
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.0
    # via google-auth
pydantic==2.7.4
    # via aiogram
pydantic-core==2.18.4
    # via pydantic
python-dotenv==1.0.1
    # via -r requirements.in
requests==2.32.3
    # via
    #   -r requirements.in
    #   requests-oauthlib
requests-oauthlib==2.0.0
    # via google-auth-oauthlib
rsa==4.9
    # via google-auth
soupsieve==2.5
    # via beautifulsoup4
typing-extensions==4.12.2
    # via
    #   aiogram
    #   pydantic
    #   pydantic-core
urllib3==2.2.2
    # via requests
yarl==1.9.4
    # via aiohttp
